setting to navigate to map page on tarkov.dev
