+++ TarkovMonitor/App.config
@@ -28,6 +28,9 @@
             <setting name="remoteId" serializeAs="String">
                 <value />
             </setting>
+            <setting name="autoNavigateMap" serializeAs="String">
+                <value>False</value>
+            </setting>
         </TarkovMonitor.Properties.Settings>
     </userSettings>
 </configuration>
