+++ TarkovMonitor/Properties/Settings.settings
@@ -23,5 +23,8 @@
     <Setting Name="remoteId" Type="System.String" Scope="User">
       <Value Profile="(Default)" />
     </Setting>
+    <Setting Name="autoNavigateMap" Type="System.Boolean" Scope="User">
+      <Value Profile="(Default)">False</Value>
+    </Setting>
   </Settings>
 </SettingsFile>
